Fixed serious bug when compiling with clang

+++ MachoLoader/Classes/InstrArgStruct.h
@@ -7,7 +7,7 @@
 //
 
 struct InstrArgStruct {
-	int numberOfArgs;
+	NSUInteger numberOfArgs;
 	void *value;
 };
